Added `vimeo/psalm`, solved hard-crash with `doctrine/dbal:^2.12`
On `doctrine/dbal:2.12.x`, a `Doctrine\DBAL\Driver\ResultStatement` is passed
to `DoctrineMessageRepository#yieldMessagesForResult()`
instead of a `Doctrine\DBAL\Driver\Statement`.

Why? For the love and glory of Satan, of course :-) (I really have no clue)

To ensure that we avoid this sort of type-level problems, we hereby add
`vimeo/psalm` to the CI pipeline with some relatively strict defaults,
and adapted the code to remove bits that are detected as potentially
problematic.

In addition to that, the implementation now returns `Generator<Message>`,
instead of `Generator<mixed>`.